refactor(visualization): move DrawContours, DrawKeypoints, DrawMatches from core to visualization
Moves the three annotation ops to improc::visualization where they
logically belong (rendering/display concerns, not core image processing).
Backward-compat using-declarations in the core shim headers and pipeline.hpp
keep all existing code compiling without changes.
